requirements.txt
crewai
crewai-tools
python-dotenv
streamlit
flask
chromadb>=0.5.10,<0.6.0
